Start cleaning up `exec/operator.hpp`
