+++ libtenzir/include/tenzir/exec/exhausted.hpp
@@ -1,0 +1,17 @@
+//    _   _____   __________
+//   | | / / _ | / __/_  __/     Visibility
+//   | |/ / __ |_\ \  / /          Across
+//   |___/_/ |_/___/ /_/       Space and Time
+//
+// SPDX-FileCopyrightText: (c) 2025 The Tenzir Contributors
+// SPDX-License-Identifier: BSD-3-Clause
+
+#pragma once
+
+namespace tenzir::exec {
+
+/// A tag type that signals within a pipeline that an operator has exhausted its
+/// inputs.
+struct exhausted {};
+
+} // namespace tenzir::exec
