+++ libtenzir/include/tenzir/exec/operator.hpp
@@ -11,21 +11,13 @@
 #include "tenzir/fwd.hpp"
 
 #include "tenzir/atoms.hpp"
-#include "tenzir/variant.hpp"
+#include "tenzir/exec/message.hpp"
 
 #include <caf/scheduled_actor/flow.hpp>
 #include <caf/typed_event_based_actor.hpp>
 #include <caf/typed_stream.hpp>
 
 namespace tenzir::exec {
-
-struct checkpoint {
-  friend auto inspect(auto& f, checkpoint& x) -> bool {
-    return f.object(x).fields();
-  }
-};
-
-struct exhausted {};
 
 // WITHOUT END
 // observable<variant<checkpoint, T>>
@@ -44,47 +36,6 @@ struct exhausted {};
 // from kafka | head 10?
 
 // DB transform operator
-
-template <class T>
-struct message : variant<checkpoint, exhausted, T> {
-  using super = variant<checkpoint, exhausted, T>;
-  using super::super;
-
-  friend auto inspect(auto& f, message<T>& x) -> bool {
-    return f.apply(static_cast<super&>(x));
-  }
-};
-
-template <>
-struct message<void> : variant<checkpoint, exhausted> {
-  using super = variant<checkpoint, exhausted>;
-  using super::super;
-
-  template <class U>
-    requires(not std::is_void_v<U>)
-  explicit(false) operator message<U>() && {
-    return as<checkpoint>(std::move(*this));
-  }
-
-  friend auto inspect(auto& f, message<void>& x) -> bool {
-    return f.apply(static_cast<super&>(x));
-  }
-};
-
-template <class T>
-using stream = caf::typed_stream<message<T>>;
-
-template <class T>
-using observable = caf::flow::observable<message<T>>;
-
-template <class T>
-struct as_stream {
-  using type = stream<T>;
-};
-
-using message_types = caf::type_list<void, table_slice>;
-using stream_types = caf::detail::tl_map_t<message_types, as_stream>;
-using any_stream = detail::tl_apply_t<stream_types, variant>;
 
 struct operator_actor_traits {
   using signatures = caf::type_list<
